+++ extras/tests/ContinuousStepperTests/begin.cpp
@@ -14,4 +14,17 @@ TEST_CASE("ContinuousStepper::begin()") {
       });
     }
   }
+
+  WHEN("begin(2, 3, 4, 5) is called") {
+    stepper.begin(2, 3, 4, 5);
+
+    THEN("it should configure the four pins as OUTPUT") {
+      CHECK_ARDUINO_LOG({
+          {0'000, "pinMode(2, OUTPUT)"},
+          {0'000, "pinMode(3, OUTPUT)"},
+          {0'000, "pinMode(4, OUTPUT)"},
+          {0'000, "pinMode(5, OUTPUT)"},
+      });
+    }
+  }
 }